debt: playground: extract balance-card + graph-inspector from main.js
Two more self-contained modules.

- main/balance-card.js: owns balanceHistory + the 48 h live points
  buffer + liveYesterdayHigh. Exposes fetchBalanceHistory,
  appendBalanceLivePoint, renderBalanceCard, and get/reset accessors
  for the yesterday-peak used by the Status view. Takes an
  onRerender callback at init time so main.js's
  rerenderWithHistoryFallback still fires after the 48h fetch
  resolves.
- main/graph-inspector.js: crosshair + tooltip on the history graph.
  Desktop hover + mobile long-press. Owns its own inspectorX state;
  reads timeSeriesStore / graphRange / showAllSensors via main.js
  live bindings.

main.js: 1992 → 1627 lines. npm test:unit 782/782 pass. knip clean.
