+++ playground/js/main.js
@@ -1,10 +1,5 @@
 import { loadSystemYaml } from './yaml-loader.js';
 import { ThermalModel, tankStoredEnergyKwh } from './physics.js';
-import {
-  computeEnergyBalance,
-  editorialDaySentence,
-  editorialNightSentence,
-} from './energy-balance.js';
 import { ControlStateMachine, initControlLogic } from './control.js';
 import { load as loadControlLogic } from './control-logic-loader.js';
 import { createSlider, formatTime, pickTickStep, formatTick, pickBucketSize } from './ui.js';
@@ -35,6 +30,11 @@ import {
   transitionLog, fetchLiveEvents, detectLiveTransition, renderLogsList,
   setupLogsScrollLoader, setupCopyLogsButton, resetEventsState,
 } from './main/logs.js';
+import {
+  initBalanceCard, fetchBalanceHistory, appendBalanceLivePoint,
+  renderBalanceCard, getLiveYesterdayHigh, resetLiveYesterdayHigh,
+} from './main/balance-card.js';
+import { setupInspector } from './main/graph-inspector.js';
 // Expose for e2e testing
 window.__triggerVersionCheck = triggerVersionCheck;
 
@@ -327,7 +327,7 @@ function clearLiveDisplay() {
   var ghTrendResetEl = document.getElementById('tank-stat-greenhouse-trend');
   if (ghTrendResetEl) ghTrendResetEl.innerHTML = '';
   document.getElementById('inactive-modes').innerHTML = '';
-  liveYesterdayHigh = null;
+  resetLiveYesterdayHigh();
   document.getElementById('graph-peak-label').textContent = "Yesterday's High: --";
   var arc = document.getElementById('tank-gauge-arc');
   if (arc) arc.setAttribute('stroke-dashoffset', '628');
@@ -357,7 +357,7 @@ function switchToSimulation() {
   balanceLivePoints = [];
   balanceLiveEvents = [];
   balanceLiveLastMode = null;
-  liveYesterdayHigh = null;
+  resetLiveYesterdayHigh();
   renderBalanceCard();
   stopStalenessTimer();
   const banner = document.getElementById('staleness-banner');
@@ -468,163 +468,8 @@ let lastDay = 0;
 // 48h /api/history response (see fetchBalanceHistory) as the peak tank
 // average across yesterday's local calendar day. null when no points
 // from yesterday are available.
-let liveYesterdayHigh = null;
+// liveYesterdayHigh moved to ./main/balance-card.js
 
-// ── Graph Inspector state ──
-let inspectorX = null; // null = hidden, otherwise CSS pixel x relative to canvas
-
-function setupInspector() {
-  const canvas = document.getElementById('chart');
-  const container = canvas.parentElement;
-  const tooltip = document.getElementById('graph-inspector');
-  const crosshair = document.getElementById('graph-crosshair');
-
-  function getCanvasX(clientX) {
-    const rect = canvas.getBoundingClientRect();
-    return clientX - rect.left;
-  }
-
-  function showInspector(x) {
-    inspectorX = x;
-    crosshair.style.display = 'block';
-    crosshair.style.left = x + 'px';
-    tooltip.style.display = 'block';
-    // Position tooltip: flip side if near right edge
-    const containerW = container.offsetWidth;
-    if (x > containerW * 0.6) {
-      tooltip.style.left = 'auto';
-      tooltip.style.right = (containerW - x + 12) + 'px';
-    } else {
-      tooltip.style.left = (x + 12) + 'px';
-      tooltip.style.right = 'auto';
-    }
-    updateInspectorData(x);
-  }
-
-  function hideInspector() {
-    inspectorX = null;
-    tooltip.style.display = 'none';
-    crosshair.style.display = 'none';
-  }
-
-  function updateInspectorData(x) {
-    if (timeSeriesStore.times.length < 2) return;
-    const dw = canvas.offsetWidth;
-    const pad = { top: 16, right: 16, bottom: 24, left: 8 };
-    const pw = dw - pad.left - pad.right;
-
-    const latestTime = timeSeriesStore.times[timeSeriesStore.times.length - 1];
-    const tMax = Math.max(graphRange, latestTime);
-    const tMin = tMax - graphRange;
-
-    // Convert pixel x to simulation time
-    const frac = (x - pad.left) / pw;
-    if (frac < 0 || frac > 1) { hideInspector(); return; }
-    const simTime = tMin + frac * graphRange;
-
-    // Find nearest data point
-    let bestIdx = 0, bestDist = Infinity;
-    for (let i = 0; i < timeSeriesStore.times.length; i++) {
-      const d = Math.abs(timeSeriesStore.times[i] - simTime);
-      if (d < bestDist) { bestDist = d; bestIdx = i; }
-    }
-
-    const v = timeSeriesStore.values[bestIdx];
-    const t = timeSeriesStore.times[bestIdx];
-
-    // Time of day — live data stores Unix epoch seconds, simulation
-    // stores seconds since sim start + SIM_START_HOUR offset.
-    let todH, todM;
-    if (store.get('phase') === 'live') {
-      const d = new Date(t * 1000);
-      todH = d.getHours();
-      todM = d.getMinutes();
-    } else {
-      todH = Math.floor((SIM_START_HOUR + t / 3600) % 24);
-      todM = Math.floor(((SIM_START_HOUR + t / 3600) % 1) * 60);
-    }
-    document.getElementById('inspector-time').textContent =
-      todH.toString().padStart(2, '0') + ':' + todM.toString().padStart(2, '0');
-
-    // Temperature values (null-tolerant for unbound live sensors)
-    const fmtInspTemp = function (x) { return isNum(x) ? x.toFixed(1) + '\u00b0C' : TEMP_PLACEHOLDER; };
-    document.getElementById('inspector-coll').textContent = fmtInspTemp(v.t_collector);
-    document.getElementById('inspector-tank').textContent = fmtInspTemp(tankAvgOf(v));
-    if (showAllSensors) {
-      document.getElementById('inspector-tank-top').textContent = fmtInspTemp(v.t_tank_top);
-      document.getElementById('inspector-tank-bottom').textContent = fmtInspTemp(v.t_tank_bottom);
-    }
-    document.getElementById('inspector-gh').textContent = fmtInspTemp(v.t_greenhouse);
-    document.getElementById('inspector-out').textContent = fmtInspTemp(v.t_outdoor);
-
-    // Duty cycle for the hour containing this point
-    const hourSeconds = 3600;
-    const hr = Math.floor(t / hourSeconds);
-    const hrStart = hr * hourSeconds;
-    const hrEnd = (hr + 1) * hourSeconds;
-    let chargingSec = 0, heatingSec = 0, emergencySec = 0, totalSec = 0;
-    for (let j = 0; j < timeSeriesStore.times.length; j++) {
-      const st = timeSeriesStore.times[j];
-      if (st >= hrStart && st < hrEnd) {
-        totalSec += 5;
-        if (timeSeriesStore.modes[j] === 'solar_charging') chargingSec += 5;
-        if (timeSeriesStore.modes[j] === 'greenhouse_heating') heatingSec += 5;
-        if (timeSeriesStore.modes[j] === 'emergency_heating') emergencySec += 5;
-      }
-    }
-    const chPct = totalSec > 0 ? Math.round(100 * chargingSec / totalSec) : 0;
-    const htPct = totalSec > 0 ? Math.round(100 * heatingSec / totalSec) : 0;
-    const emPct = totalSec > 0 ? Math.round(100 * emergencySec / totalSec) : 0;
-    document.getElementById('inspector-charging').textContent = chPct + '%';
-    document.getElementById('inspector-heating').textContent = htPct + '%';
-    document.getElementById('inspector-emergency').textContent = emPct + '%';
-  }
-
-  // Desktop: hover
-  canvas.addEventListener('mousemove', function(e) {
-    showInspector(getCanvasX(e.clientX));
-  });
-  canvas.addEventListener('mouseleave', hideInspector);
-
-  // Mobile: long press
-  let longPressTimer = null;
-  let longPressActive = false;
-
-  canvas.addEventListener('touchstart', function(e) {
-    if (e.touches.length !== 1) return;
-    const touch = e.touches[0];
-    const startX = touch.clientX;
-    const startY = touch.clientY;
-    longPressTimer = setTimeout(function() {
-      longPressActive = true;
-      showInspector(getCanvasX(startX));
-    }, 400);
-  }, { passive: true });
-
-  canvas.addEventListener('touchmove', function(e) {
-    if (longPressActive) {
-      e.preventDefault();
-      showInspector(getCanvasX(e.touches[0].clientX));
-    } else {
-      // User is scrolling, cancel long press
-      clearTimeout(longPressTimer);
-    }
-  }, { passive: false });
-
-  canvas.addEventListener('touchend', function() {
-    clearTimeout(longPressTimer);
-    if (longPressActive) {
-      longPressActive = false;
-      hideInspector();
-    }
-  });
-
-  canvas.addEventListener('touchcancel', function() {
-    clearTimeout(longPressTimer);
-    longPressActive = false;
-    hideInspector();
-  });
-}
 
 // ── Time Series Store (extended with mode tracking) ──
 export const timeSeriesStore = {
@@ -807,6 +652,7 @@ async function init() {
   setupInspector();
   setupLogsScrollLoader();
   setupCopyLogsButton();
+  initBalanceCard({ onRerender: rerenderWithHistoryFallback });
   updateDisplay(model.getState(), { mode: 'idle', valves: { vi_btm: false, vi_top: false, vi_coll: false, vo_coll: false, vo_rad: false, vo_tank: false, v_air: false }, actuators: { pump: false, fan: false, space_heater: false }, transition: null });
 
   // Initialize live/simulation mode toggle
@@ -924,217 +770,6 @@ function fetchLiveHistory(rangeSeconds) {
     .catch(() => { liveHistoryData = null; });
 }
 
-// ── Today's balance card ──────────────────────────────────────────────
-// The balance card on the Status view needs a fixed 48 h slice that
-// doesn't shrink when the user flips the history graph's 1 H/6 H/… pill.
-// We pull it separately and apply live state frames on top so the numbers
-// stay fresh without hammering /api/history on every WebSocket message.
-//
-// balanceHistory holds the raw server response; balanceLivePoints is an
-// incremental tail of points appended from live WS frames (rate-limited
-// to one per 5 min, matching the database aggregation cadence).
-let balanceHistory = null;
-let balanceLivePoints = [];
-let balanceLiveEvents = [];
-let balanceLiveLastMode = null;
-let balanceLastAppendTs = 0;
-const BALANCE_APPEND_MIN_INTERVAL_MS = 5 * 60 * 1000;
-
-function fetchBalanceHistory() {
-  if (store.get('phase') !== 'live') return;
-  fetch('/api/history?range=48h')
-    .then(r => r.json())
-    .then(data => {
-      if (store.get('phase') !== 'live') return;
-      balanceHistory = data;
-      balanceLivePoints = [];
-      balanceLiveEvents = [];
-      balanceLiveLastMode = null;
-      liveYesterdayHigh = computeLiveYesterdayHigh(data && data.points);
-      renderBalanceCard();
-      // Balance fetch completes independently of the WS state stream,
-      // so re-render so the peak label reflects the freshly-computed
-      // yesterdayHigh instead of waiting for the next state frame.
-      rerenderWithHistoryFallback();
-    })
-    .catch(() => { balanceHistory = null; });
-}
-
-// Peak tank average ((tank_top + tank_bottom) / 2) across points whose
-// timestamps fall within yesterday's local calendar day. Matches what
-// the graph's Tank line plots and the central tank gauge displays.
-// Returns null when no qualifying points exist.
-function computeLiveYesterdayHigh(points) {
-  if (!Array.isArray(points) || points.length === 0) return null;
-  const now = new Date();
-  const yStart = new Date(now.getFullYear(), now.getMonth(), now.getDate() - 1).getTime();
-  const yEnd = new Date(now.getFullYear(), now.getMonth(), now.getDate()).getTime();
-  let peak = null;
-  for (let i = 0; i < points.length; i++) {
-    const p = points[i];
-    if (!p || typeof p.ts !== 'number') continue;
-    if (p.ts < yStart || p.ts >= yEnd) continue;
-    if (!isNum(p.tank_top) || !isNum(p.tank_bottom)) continue;
-    const avg = (p.tank_top + p.tank_bottom) / 2;
-    if (peak === null || avg > peak) peak = avg;
-  }
-  return peak;
-}
-
-function appendBalanceLivePoint(state, result) {
-  if (store.get('phase') !== 'live') return;
-  if (!isNum(state.t_tank_top) || !isNum(state.t_tank_bottom)) return;
-
-  // Track mode changes as synthetic events so computeEnergyBalance
-  // classifies recent deltas against the right mode (otherwise every
-  // post-fetch sample would inherit the last historical mode).
-  const currentMode = (result && result.mode) ? String(result.mode).toLowerCase() : null;
-  if (currentMode && currentMode !== balanceLiveLastMode) {
-    balanceLiveEvents.push({ ts: Date.now(), type: 'mode', to: currentMode });
-    balanceLiveLastMode = currentMode;
-  }
-
-  const nowMs = Date.now();
-  if (nowMs - balanceLastAppendTs < BALANCE_APPEND_MIN_INTERVAL_MS) return;
-  balanceLastAppendTs = nowMs;
-  balanceLivePoints.push({
-    ts: nowMs,
-    tank_top: state.t_tank_top,
-    tank_bottom: state.t_tank_bottom,
-  });
-  renderBalanceCard();
-}
-
-function fmtBalanceKwh(v, { sign = false } = {}) {
-  const rounded = Math.round(v * 10) / 10;
-  if (sign) {
-    if (rounded === 0) return '0.0';
-    return (rounded > 0 ? '+' : '−') + Math.abs(rounded).toFixed(1);
-  }
-  return rounded.toFixed(1);
-}
-
-function fmtBalanceWindow(startTs, endTs, complete) {
-  const now = new Date();
-  const fmt = (ts, refDate) => {
-    const d = new Date(ts);
-    const sameDay = d.toDateString() === refDate.toDateString();
-    const hm = d.toLocaleTimeString('en-GB', { hour: '2-digit', minute: '2-digit' });
-    if (sameDay) return hm;
-    const yesterday = new Date(refDate);
-    yesterday.setDate(refDate.getDate() - 1);
-    if (d.toDateString() === yesterday.toDateString()) return hm + ' yesterday';
-    return hm + ' ' + d.toLocaleDateString('en-GB', { day: 'numeric', month: 'short' });
-  };
-  return fmt(startTs, now) + ' → ' + (complete ? fmt(endTs, now) : 'now');
-}
-
-function statHtml(label, kwh, { sign = false, extra = '' } = {}) {
-  const val = fmtBalanceKwh(kwh, { sign });
-  const cls = sign
-    ? (kwh >= 0 ? ' positive' : ' negative')
-    : '';
-  const capHtml = extra ? '<span class="balance-stat-caption">' + extra + '</span>' : '';
-  return '<div class="balance-stat">' +
-    '<span class="balance-stat-label">' + label + '</span>' +
-    '<div><span class="balance-stat-value' + cls + '">' + val + '</span>' +
-    '<span class="balance-stat-unit">kWh</span></div>' +
-    capHtml +
-    '</div>';
-}
-
-function releasedStatHtml(heatingKwh, leakageKwh) {
-  const total = heatingKwh + leakageKwh;
-  const heating = heatingKwh >= 0.05;
-  const leakage = leakageKwh >= 0.05;
-  let caption;
-  if (heating && leakage) {
-    caption = fmtBalanceKwh(heatingKwh) + ' to greenhouse · ' +
-      fmtBalanceKwh(leakageKwh) + ' to air';
-  } else if (heating) {
-    caption = 'to greenhouse';
-  } else if (leakage) {
-    caption = 'to air';
-  } else {
-    caption = '';
-  }
-  return '<div class="balance-stat">' +
-    '<span class="balance-stat-label">Released</span>' +
-    '<div><span class="balance-stat-value">−' + fmtBalanceKwh(total) + '</span>' +
-    '<span class="balance-stat-unit">kWh</span></div>' +
-    (caption ? '<span class="balance-stat-caption">' + caption + '</span>' : '') +
-    '</div>';
-}
-
-function renderBalanceSection(section, window, sentence, stats) {
-  if (!stats) {
-    section.hidden = true;
-    return;
-  }
-  section.hidden = false;
-  section.querySelector('[id$="-window"]').textContent = window;
-  section.querySelector('[id$="-sentence"]').textContent = sentence;
-  section.querySelector('[id$="-stats"]').innerHTML = stats;
-}
-
-function renderBalanceCard() {
-  const card = document.getElementById('balance-card');
-  if (!card) return;
-  if (store.get('phase') !== 'live' || !balanceHistory) {
-    card.style.display = 'none';
-    return;
-  }
-  const points = (balanceHistory.points || []).concat(balanceLivePoints);
-  const events = (balanceHistory.events || []).concat(balanceLiveEvents);
-  if (points.length < 2) {
-    card.style.display = 'none';
-    return;
-  }
-
-  const balance = computeEnergyBalance(points, events, Date.now());
-  if (!balance.night && !balance.day) {
-    card.style.display = 'none';
-    return;
-  }
-  card.style.display = '';
-
-  // Night section
-  const nightSec = document.getElementById('balance-night');
-  if (balance.night) {
-    const n = balance.night;
-    const hasLoss = (n.heatingKwh + n.leakageKwh) >= 0.05;
-    const stats = hasLoss
-      ? releasedStatHtml(n.heatingKwh, n.leakageKwh)
-      : statHtml('Net', n.netKwh, { sign: true });
-    renderBalanceSection(
-      nightSec,
-      (n.complete ? 'Night · ' : 'Night so far · ') + fmtBalanceWindow(n.startTs, n.endTs, n.complete),
-      editorialNightSentence(n),
-      stats,
-    );
-  } else {
-    nightSec.hidden = true;
-  }
-
-  // Day section
-  const daySec = document.getElementById('balance-day');
-  if (balance.day) {
-    const d = balance.day;
-    const stats = [
-      statHtml('Gathered', d.gatheredKwh),
-      releasedStatHtml(d.heatingKwh, d.leakageKwh),
-      statHtml('Net today', d.netKwh, { sign: true }),
-    ].join('');
-    renderBalanceSection(
-      daySec,
-      (d.complete ? 'Day · ' : 'Day so far · ') + fmtBalanceWindow(d.startTs, d.endTs, d.complete),
-      editorialDaySentence(d),
-      stats,
-    );
-  } else {
-    daySec.hidden = true;
-  }
-}
 
 // Re-render the status/components views after something refills the
 // timeSeriesStore (currently just the live-history fetch). Uses the most
@@ -1802,7 +1437,7 @@ function updateDisplay(state, result) {
   // Graph yesterday's high label. Simulation tracks a per-sim-day peak
   // via confirmedYesterdayHigh; live mode derives it from the 48h
   // history fetch (liveYesterdayHigh) since state.simTime does not tick.
-  const peakVal = store.get('phase') === 'live' ? liveYesterdayHigh : confirmedYesterdayHigh;
+  const peakVal = store.get('phase') === 'live' ? getLiveYesterdayHigh() : confirmedYesterdayHigh;
   document.getElementById('graph-peak-label').textContent =
     isNum(peakVal) && peakVal > 0 ? `Yesterday's High: ${peakVal.toFixed(0)}°C` : 'Yesterday\'s High: --';
